monitoring/exporters/bigquery-exporter/requirements.txt
google-cloud-bigquery==3.11.4
prometheus-client==0.19.0
flask==2.3.3
gunicorn==21.2.0
